refactor(sw): extract tab affordances to background/tab-affordances.js
Three cohesive tab-strip affordances lived inline in the SW as two blocks
plus scattered listeners: the agent-tab card ("peerd opened <tab>",
anchored to the driving message_actor turn), the "pull peerd in" web-tab
hint (an informational, auto-dismissing injected caption), and the
toolbar-icon / Alt+Shift+P front door. All pure UI presence, no security
surface, sharing only tab-strip state.

Move them into one helper module that owns every reassigned bit of that
state (agentTab*, activeTabId, peerdWebTabs, homeTabIds,
lastFocusedWindowId, tabMsgAnchor) and registers its own listeners. The SW
destructures the six methods it still calls (noteAgentTab,
broadcastAgentTab, scheduleWebTabHint, showWebTabHint, isHomeOpen, and
setTabAnchor for the actor-turn start, which replaces the direct
tabMsgAnchor.set). The sync-inside-the-gesture constraint on pullInPeerd
is preserved verbatim.

The pull-in hint reads our own bundled icon32.png via runtime.getURL, so
the module joins the system-prompt loader's no-restricted-globals
exemption (service-worker.js had that rule off). No behavior change.
ts-check floor 483 → 484.

Verified: 2545 Bun tests, typecheck, ESLint, 596 in-browser tests, and the
full 17-state / 94-check e2e (actor-delegate exercises the agent-tab card
path); side-panel render confirmed by screenshot.
