+++ extension/background/service-worker.js
@@ -30,7 +30,6 @@
 import browser from '/vendor/browser-polyfill.js';
 import { makeDispatcher, isTrustedSender } from '/shared/messaging.js';
 import { CHANNEL_DEFAULTS, CHANNEL, DWEB_ENABLED } from '/shared/channel-config.js';
-import { openHome } from '/shared/open-home.js';
 import { REMOTE_SKILL_INSTALL } from '/shared/flags.js';
 
 import {
@@ -237,8 +236,6 @@ import {
   // (addressing + same-origin-lock anchor), and the "what I learned" self-fence.
   makeApiActorBindings, normalizeApiOrigin, fenceApiActorSummary,
   finalAssistantText,
-  // The informational "pull peerd in" reminder injected into peerd-opened web tabs.
-  pullInHintInjected,
 } from '/peerd-runtime/index.js';
 
 import { flattenCategorisedDenylist, normalizeDenylistPattern } from '/peerd-egress/index.js';
@@ -251,7 +248,6 @@ import { makeOffscreenJsClient } from './offscreen-js-client.js';
 import { makeOffscreenActorClient } from './offscreen-actor-client.js';
 import { makeOffscreenPdfClient } from './offscreen-pdf-client.js';
 import { makeUiPorts } from './ui-ports.js';
-import { decidePullIn } from './panel-affordance.js';
 import { createAppClient, APP_TAB_GROUP_TITLE } from './app-client.js';
 import { createAppTabTracker } from './app-tab-tracker.js';
 import {
@@ -285,6 +281,7 @@ import { makeSessionState } from './session-state.js';
 import { makeLocalModelState } from './local-model-state.js';
 import { makeProfileState } from './profile-state.js';
 import { makeModelCatalog } from './model-catalog.js';
+import { makeTabAffordances } from './tab-affordances.js';
 import { makeVaultRoutes } from './routes/vault.js';
 import { makeProviderRoutes } from './routes/providers.js';
 import { makeHooksRoutes } from './routes/hooks.js';
@@ -1886,6 +1883,15 @@ const closeSidePanel = async () => {
   }
 };
 
+// How peerd shows up in the tab strip (background/tab-affordances.js): the
+// agent-tab card, the "pull peerd in" web-tab hint, and the toolbar-icon /
+// Alt+Shift+P front door. It owns all the tab-strip state + listeners; the SW
+// calls noteAgentTab/scheduleWebTabHint/broadcastAgentTab/showWebTabHint from
+// its tool-context + port wiring, and setTabAnchor from the actor-turn start.
+const {
+  noteAgentTab, broadcastAgentTab, scheduleWebTabHint, showWebTabHint, setTabAnchor, isHomeOpen,
+} = makeTabAffordances({ browser, uiPorts, denylistStore, closeSidePanel });
+
 // Confirmation coordinator. The dispatcher's async confirmation step
 // calls ctx.confirm(prompt); this pushes a 'confirm/request' to the side
 // panel and resolves when the panel posts back 'confirm/answer'.
@@ -3120,7 +3126,7 @@ const actorMessaging = makeActorMessaging({
         : kind === 'notebook' ? jsTabTracker.getTabId(instanceId)
         : kind === 'app' ? appTabTracker.getTabId(instanceId)
         : null;
-      if (typeof ownedTab === 'number') tabMsgAnchor.set(ownedTab, parentToolUseId);
+      if (typeof ownedTab === 'number') setTabAnchor(ownedTab, parentToolUseId);
     }
     // DESIGN-17 P1 glass pane: when this turn was triggered by a live message_actor
     // call (parentToolUseId present — absent on a boot redrain), pass a `display`
@@ -3204,147 +3210,6 @@ const postChatNote = (/** @type {string} */ text, /** @type {any} */ action = nu
   try { uiPorts.broadcast({ type: 'turn/system-note', text, ...(action ? { action } : {}) }); }
   catch { /* panel gone */ }
 };
-
-// The CURRENT AGENT TAB — the single tab the loop most recently created OR
-// interacted with (ran a command in, navigated, clicked, …). Agent tabs open in
-// the BACKGROUND now (never steal focus), so the chat shows ONE persistent,
-// sticky "go to the tab peerd is working in" card pointing here; clicking it is
-// the user gesture that focuses the tab AND opens the side panel (Chrome won't
-// let the agent/SW open the panel on its own — DESIGN-12). Updated on every
-// touch, so it always tracks where the agent IS, not just the last tab created.
-// p·cyan e·red e·amber r·green d·magenta — the home agent-tab card's Open button
-// draws a fresh one each time the card is (re)generated.
-const AGENT_TAB_COLORS = ['#00B7EB', '#EF4444', '#F59E0B', '#22C55E', '#D946EF'];
-// DESIGN-17/18 tab-card anchoring: maps an agent tab → the message_actor tool_use that
-// LAST drove it. The inline "peerd opened …" notice anchors to THAT message's turn, not
-// the wall-clock-latest user message — actor work is async, so a physical tab touch
-// (engine ensureTab / web DOM noteTab) often lands during a later turn, which would clump
-// the cards at the chat's end. Set at each actor-turn start (runActorTurn) for the actor's
-// owned tab; never cleared — overwritten only when a NEW message re-drives that tab, which
-// is exactly when the card should resurface to the newer turn. Orchestrator-opened tabs
-// (open_tab) are absent here → they keep the wall-clock anchor (correct; they're synchronous).
-/** @type {Map<number, string>} tabId → parentToolUseId */
-const tabMsgAnchor = new Map();
-/** @type {number | null} */ let agentTabId = null;
-/** @type {any} */ let agentTabInfo = null;   // the last { tabId, windowId, kind, name, label, color } noted
-/** @type {number | null} */ let activeTabId = null;    // the currently-active tab — hide the card when you're ON it
-// Broadcast the current-agent-tab pointer. `noted` is true ONLY when this fires
-// from a real agent touch (noteAgentTab) — the inline notice creates/resurfaces
-// on those; a passive refresh (tab activation, a fresh surface replay) sends
-// noted:false so clicking around tabs never bumps a notice.
-const broadcastAgentTab = (noted = false) => {
-  uiPorts.broadcast({
-    type: 'agent/tab',
-    tab: agentTabInfo ? { ...agentTabInfo, current: agentTabInfo.tabId === activeTabId, noted } : null,
-  });
-};
-const noteAgentTab = async (/** @type {number} */ tabId, /** @type {any} */ info = {}) => {
-  if (typeof tabId !== 'number') return;
-  const { kind = null, name = null, label = null, opened = true, parentToolUseId: ptuArg = null } = (typeof info === 'string' ? { label: info } : info);
-  // The message_actor turn driving this tab (see tabMsgAnchor) — caller-supplied or the
-  // last actor-turn-start mapping. Flows to the agent/tab event so the notice anchors to
-  // that message's turn instead of the wall-clock-latest user message.
-  const parentToolUseId = ptuArg ?? tabMsgAnchor.get(tabId) ?? null;
-  let windowId; let title = null;
-  try { const t = await browser.tabs.get(tabId); windowId = t.windowId; title = t.title || t.url || null; }
-  catch { return; } // tab already gone — don't point the card at a dead tab
-  agentTabId = tabId;
-  // Instance tabs carry a kind + the instance NAME (the card reads like a tab:
-  // "Notebook | my-nb"); a web tab (open_tab / DOM) just shows its page label.
-  const text = (kind && name) ? `${kind} · ${name}` : (label || title || 'a tab');
-  // why a fresh brand color each generation (owner): the home card's Open button
-  // cycles a peerd brand color (p·cyan e·red e·amber r·green d·magenta) so it
-  // stays eye-catching — the sanctioned "peers/actions are the content" accent.
-  const color = AGENT_TAB_COLORS[Math.floor(Math.random() * AGENT_TAB_COLORS.length)];
-  // `opened`: true when peerd OPENED this tab (open_tab / an engine create) — the
-  // only case that mints an inline notice; false when the web actor merely ACTED
-  // on a tab, which resurfaces an existing notice but never invents one for a tab
-  // the USER opened. `noted: true` marks this as a real agent touch (vs. a
-  // passive current-flag refresh on tab activation).
-  agentTabInfo = { tabId, windowId, kind, name, label: text, color, opened, parentToolUseId };
-  broadcastAgentTab(true);
-};
-// Track the active tab so the card hides when you're on the agent tab, and shows
-// again when you move away.
-browser.tabs?.onActivated?.addListener(({ tabId }) => {
-  activeTabId = tabId;
-  if (agentTabInfo) broadcastAgentTab();
-});
-// Clear the card when the agent tab closes (clicking a dead tab does nothing).
-browser.tabs?.onRemoved?.addListener((tabId) => {
-  if (tabId === agentTabId) { agentTabId = null; agentTabInfo = null; broadcastAgentTab(); }
-});
-
-// "Pull peerd in" reminder on the regular web pages peerd opens. A peerd-opened
-// WEB tab gets a brief, auto-dismissing caption (top-right) that types out
-// "Press <shortcut> to pull peerd in" — engine tabs carry the real button, a
-// third-party page can't, so this points you at the shortcut/icon. INFORMATIONAL
-// ONLY (it never messages the SW back), so it crosses no boundary and needs no
-// new permission (docs/PULL-IN-PEERD-WEB-SCOPE.md). One-shot, on first load, via
-// chrome.scripting; never on a denylisted/sensitive origin; the injected script
-// itself waits until the tab is actually visible before it shows.
-// Peerd-opened web tabs (tabId → origin), tracked persistently so we can show
-// the reminder at the RIGHT moment: when the user is ACTIVELY VIEWING one with
-// the SIDEBAR CLOSED — they walked onto it, OR they closed the panel while on it.
-// The page world can't read sidebar state, so the SW gates the inject; the
-// injected script is idempotent + auto-dismissing, so re-injecting is safe.
-/** @type {Map<number, string>} */
-const peerdWebTabs = new Map();
-// The peerd toolbar icon as a data: URL, so the injected hint can show it on a
-// third-party page without a chrome-extension:// fetch (no web_accessible_
-// resources needed). Fetched + cached once; '' if it ever fails (the hint then
-// falls back to the wordmark text).
-/** @type {string | null} */ let pullInIconUrl = null;
-const getPullInIconUrl = async () => {
-  if (pullInIconUrl !== null) return pullInIconUrl;
-  try {
-    const res = await fetch(browser.runtime.getURL('icons/icon32.png'));
-    const bytes = new Uint8Array(await res.arrayBuffer());
-    pullInIconUrl = `data:image/png;base64,${btoa(String.fromCharCode(...bytes))}`;
-  } catch (e) {
-    console.debug('[sw] pull-in icon load failed', (/** @type {{ message?: string }} */ (e))?.message ?? e);
-    pullInIconUrl = '';
-  }
-  return pullInIconUrl;
-};
-const showWebTabHint = async (/** @type {number} */ tabId) => {
-  if (!peerdWebTabs.has(tabId)) return;
-  if (uiPorts.hasNamed('sidepanel')) return;          // sidebar open → the chat's already here
-  let tab;
-  try { tab = await browser.tabs.get(tabId); } catch { return; }
-  if (!tab || tab.status !== 'complete' || tab.active !== true) return; // only when actually being viewed
-  // Still the page peerd opened? (don't graffiti the user's own later navigation.)
-  let origin;
-  try { origin = new URL(/** @type {string} */ (tab.url)).origin; } catch { return; }
-  if (origin !== peerdWebTabs.get(tabId)) { peerdWebTabs.delete(tabId); return; }
-  let shortcut = '';
-  try {
-    const cmds = await browser.commands?.getAll?.();
-    shortcut = (cmds ?? []).find((c) => c.name === 'pull-in-peerd')?.shortcut || '';
-  } catch { /* no commands API in this build */ }
-  const iconUrl = await getPullInIconUrl();
-  try {
-    await browser.scripting.executeScript({ target: { tabId }, func: pullInHintInjected, args: [shortcut, iconUrl] });
-  } catch (e) {
-    // Pages the browser refuses to inject into (chrome:, the stores, a hard CSP)
-    // — harmless; the hint just doesn't show.
-    console.debug('[sw] pull-in hint inject skipped', (/** @type {{ message?: string }} */ (e))?.message ?? e);
-  }
-};
-const scheduleWebTabHint = (/** @type {number} */ tabId, /** @type {string} */ url) => {
-  if (typeof tabId !== 'number' || typeof url !== 'string') return;
-  let u;
-  try { u = new URL(url); } catch { return; } // not a real web URL → no hint
-  if (!u.protocol.startsWith('http')) return;
-  if (matchesDenylist(u.hostname, denylistStore.patterns())) return; // never graffiti a sensitive site
-  peerdWebTabs.set(tabId, u.origin);
-  showWebTabHint(tabId); // if the user is already viewing it with the sidebar closed
-};
-// Show when the user WALKS ONTO a peerd web tab, or it finishes loading while
-// they're on it. (Sidebar-close is handled at the port disconnect, below.)
-browser.tabs?.onActivated?.addListener(({ tabId }) => { showWebTabHint(tabId); });
-browser.tabs?.onUpdated?.addListener((tabId, changeInfo) => { if (changeInfo.status === 'complete') showWebTabHint(tabId); });
-browser.tabs?.onRemoved?.addListener((tabId) => { peerdWebTabs.delete(tabId); });
 
 // /init orchestration lives in peerd-runtime/memory/init-orchestrator.js
 // (scan → draft → confirm → persist); the SW binds the IO. The
@@ -3597,98 +3462,9 @@ browser.runtime.onMessage.addListener(/** @type {any} */ (makeDispatcher({
   ...(/** @type {any} */ (originCredentialRoutes)),
 })));
 
-// ---------------------------------------------------------------------------
-// 7. Toolbar icon + "pull in peerd" shortcut → home, or pull the chat to the side
-// ---------------------------------------------------------------------------
-
-// The toolbar icon is peerd's FRONT DOOR. With no home up yet it opens the
-// full-page home — peerd should feel first-party, not a bolted-on sidebar
-// (DESIGN-12, owner 2026-06-20). Once home IS up, the icon COMPLEMENTS: it pulls
-// the chat into the window-global side panel (Chrome) / sidebar (Firefox) so the
-// chat follows you onto ANY tab — including a plain web page peerd opened, the
-// case the engine-tab "pull in peerd" button couldn't reach without breaching the
-// fail-closed SW boundary (docs/PULL-IN-PEERD-WEB-SCOPE.md). The Alt+Shift+P
-// command is the dedicated twin: it ALWAYS pulls the panel in, from anywhere.
-//
-// Hard constraint: sidePanel.open()/sidebarAction.open() must run SYNCHRONOUSLY
-// inside the click/keystroke gesture — no await before them or the activation is
-// dropped. So every decision input must be available without awaiting: the
-// window id (from the listener's tab arg) and "is home open?" (two sync signals
-// below). We cannot tabs.query() in the gesture; decidePullIn is a pure sync fn.
-
-// Sync "is home open?" — a boot-seeded set of home tab ids OR a live home port.
-// why both: the set survives an SW respawn (the home port may not have
-// reconnected in the instant the icon fires); the port covers a home tab the set
-// hasn't learned yet. A miss is benign — openHome() is focus-or-create, so the
-// worst case is the first post-respawn click focusing home instead of the panel.
-const HOME_URL = browser.runtime.getURL('home/home.html');
-/** @type {Set<number>} */
-const homeTabIds = new Set();
-const trackHomeTab = (/** @type {number} */ tabId, /** @type {string} */ url) => {
-  if (typeof url !== 'string') return;
-  if (url.startsWith(HOME_URL)) homeTabIds.add(tabId);
-  else homeTabIds.delete(tabId);
-};
-browser.tabs?.query?.({}).then((tabs) => {
-  for (const t of tabs) if (t.id != null) trackHomeTab(t.id, t.url ?? '');
-}).catch((e) => console.debug('[sw] home-tab bootstrap failed', e));
-// A second onUpdated/onRemoved pair (the DOM-ref ones live earlier) — keeping the
-// home-tab bookkeeping self-contained here reads cleaner than threading it in.
-browser.tabs?.onUpdated?.addListener((tabId, changeInfo, tab) => {
-  if (changeInfo.url != null || tab?.url != null) trackHomeTab(tabId, /** @type {string} */ (changeInfo.url ?? tab.url));
-});
-browser.tabs?.onRemoved?.addListener((tabId) => { homeTabIds.delete(tabId); });
-const isHomeOpen = () => homeTabIds.size > 0 || uiPorts.hasNamed('home');
-
-// A synchronous current-window id for sidePanel.open({ windowId }). onClicked and
-// (modern) onCommand both supply the tab, so this only backstops engines whose
-// command callback omits it. Seeded at boot, kept warm on focus changes.
-/** @type {number | null} */ let lastFocusedWindowId = null;
-browser.windows?.getLastFocused?.().then((w) => { lastFocusedWindowId = w?.id ?? lastFocusedWindowId; }).catch(() => {});
-browser.windows?.onFocusChanged?.addListener((winId) => {
-  if (winId != null && winId !== browser.windows.WINDOW_ID_NONE) lastFocusedWindowId = winId;
-});
-
-// The pull-in itself — open() runs synchronously (no await before it) to keep the
-// gesture; a failed/declined open falls back to home so the icon never dead-ends.
-const pullInPeerd = (/** @type {number} */ windowId, { fromShortcut = false } = {}) => {
-  const target = decidePullIn({
-    homeOpen: isHomeOpen(),
-    panelOpen: uiPorts.hasNamed('sidepanel'),
-    hasSidePanel: !!(/** @type {any} */ (browser)).sidePanel?.open,
-    hasSidebar: !!browser.sidebarAction?.open,
-    fromShortcut,
-  });
-  // Toggle-closed (shortcut only) — close needs no gesture, so fire and forget.
-  if (target === 'close') { closeSidePanel(); return; }
-  try {
-    if (target === 'panel' && windowId != null) {
-      const p = (/** @type {any} */ (browser)).sidePanel.open({ windowId });
-      if (p?.catch) p.catch((/** @type {any} */ e) => { console.warn('[sw] sidePanel.open failed', e); openHome(); });
-      return;
-    }
-    if (target === 'sidebar') {
-      const p = browser.sidebarAction.open();
-      if (p?.catch) p.catch((e) => { console.warn('[sw] sidebarAction.open failed', e); openHome(); });
-      return;
-    }
-  } catch (e) { console.warn('[sw] pull-in open threw', e); }
-  openHome();
-};
-
-browser.action?.onClicked?.addListener((tab) => {
-  pullInPeerd(/** @type {number} */ (tab?.windowId ?? lastFocusedWindowId), { fromShortcut: false });
-});
-// Alt+Shift+P (user-rebindable at the browser's extension-shortcuts page) —
-// TOGGLES the panel: pulls it in, or closes it if already open. The command
-// handler is a VALID user-gesture context on BOTH Chrome and Firefox, so it needs
-// no content-script relay — and thus no hole in the fail-closed SW boundary the
-// injected-web-page button would have required.
-browser.commands?.onCommand?.addListener((command, tab) => {
-  if (command !== 'pull-in-peerd') return;
-  pullInPeerd(/** @type {number} */ (tab?.windowId ?? lastFocusedWindowId), { fromShortcut: true });
-});
-
+// The toolbar icon + Alt+Shift+P front door (open home, or pull the chat panel
+// in) lives in background/tab-affordances.js alongside the agent-tab card and
+// web-tab hint — it owns the sync-gesture pull-in and its listeners.
 loadUserEndpoints();
 loadSettings();
